Uploading FEWD homework 2 (Monument blog)
A few questions I jotted down during the homework: 

I added my google fonts import @import to the css file instead of the Html file. is that okay?

when you're looking at a website mock, what is the edge of the page? 

fixing the capitalization issue in the text body - didn't manage to get to that. Is there a good solution? Making everything uppercase in the "basic version" of the assignment created another separate issue when I tried to get all the lorem ipsum body text to go back to capitalizing the start of each sentence. 

how would i have centered the medium image if I wanted to? The assignment doesn't call for it, but I couldn't figure out how to do it. Is an html inlign center command not appropriate?

I'm struggling a bit with being systematic when it comes to applying margins/padding to bottoms, tops etc. By the end of the medium part of the challenge i was just adding space in an ad-hoc fashion.