Add files via upload
Hey, i've some questions on these bonus task. Will come earlier to class to find out. Thanks.

Add a hover effect to anchors, buttons, social media icons. 
Create a fixed background effect on scroll for header image
Turn the nav into a "sticky nav" so that it stays fixed while the user scrolls.
Add an effect to the images in the gallery on hover.